requirements.txt
google-generativeai
twilio
Flask
gunicorn